scpmodel-c-team-mobcomp/images/ue-manager/requirements.txt
click
pyyaml
docker